refactor: centralize runtime contract adapters
